+++ src/main/java/com/igearfs/jnlp/security/TrustStoreManager.java
@@ -35,7 +35,7 @@ public class TrustStoreManager
     public static X509Certificate downloadCertificate(String jnlpUrl) throws Exception
     {
         URL url = new URL(jnlpUrl);
-        System.out.println("URL made");
+//        System.out.println("URL made");
 
         // Use a custom TrustManager that doesn't perform any certificate verification
         TrustManager[] trustAllCertificates = new TrustManager[]{
@@ -171,7 +171,7 @@ public class TrustStoreManager
             System.out.println("Cert Downloaded");
 
             // Step 3: Save the downloaded certificate to a file
-            String certFilePath = "server-cert.cer";
+            String certFilePath = entry.getName() + "_" + entry.getId();
             saveCertificateToFile(cert, certFilePath);
 
             // Step 4: Add the certificate to the JRE truststore using keytool