+++ RELEASE_NOTES.md
@@ -14,3 +14,13 @@ v 1.4.1
 v 1.5.1
 * Added clear cache button so you can download new jar's and update
 * Added the ability to add in username/password for auto login when you launch.
+
+V 1.6.1
+Fixing build issue for each os. Have windows and mac. Next Linux
+
+v 1.7.1
+Adds webstart if it is not at the end of the url.
+Icon defaults to Rocket if not selected
+Fixed UI order on right pane and creation.
+Cancel button removed from save confirmation.
+System printouts reduced to logger.debug and logger.error. (not all of them)